refactor to use built-in validators, split into a separate file
