+++ validators/validators.utils.js
@@ -1,0 +1,3 @@
+exports.getErrorMessageForMissingData = (param) => {
+  return { status: 400, msg: `Incomplete data provided: missing ${param}` };
+};
